refactor: rename update password catalog path command name
Refs: #41
