+++ src/Core/Pango.Application/UseCases/Password/Commands/MovePasswordsToCatalog/MovePasswordsToCatalogCommand.cs
@@ -1,0 +1,14 @@
+﻿using ErrorOr;
+using MediatR;
+
+namespace Pango.Application.UseCases.Password.Commands.MovePasswordsToCatalog;
+
+public class MovePasswordsToCatalogCommand : IRequest<ErrorOr<bool>>
+{
+    public MovePasswordsToCatalogCommand(Dictionary<Guid, string> passwordIdCatalogPathPairs)
+    {
+        PasswordIdCatalogPathPairs = passwordIdCatalogPathPairs;
+    }
+
+    public Dictionary<Guid, string> PasswordIdCatalogPathPairs { get; }
+}
